feat: Refactor checkout logic by moving utility functions to utils.js
